@defer directive complete
The defer directive instructs the executor to not wait for the result of
the deferred field, fragment, or inline fragment spread before sending
the initial response. Resolved deferred fields will be sent in separate
follow-up responses, which specify the path to the updated field along
with the data.

Note: if a resolver returns a result before the initial response has
been sent, GraphQL-Go takes the opportunity to include the field in the
initial response despite the field being deferred. As a result, deferred
fields may sometimes come back with the initial response, if other
non-deferred fields took longer to execute, holding the initial response
long enough for a result.